+++ Doom3/Source/Engine/Rendering/Pipeline/RenderingCullingManager.cpp
@@ -17,7 +17,7 @@ dooms::graphics::RenderingCullingManager::RenderingCullingManager()
 void dooms::graphics::RenderingCullingManager::Initialize()
 {
 	mCullingSystem = std::make_unique<culling::EveryCulling>(graphicsAPISetting::GetScreenWidth(), graphicsAPISetting::GetScreenHeight());
-	mCullingSystem->mMaskedSWOcclusionCulling->mSolveMeshRoleStage.mOccluderAABBScreenSpaceMinArea = ConfigData::GetSingleton()->GetConfigData().GetValue<FLOAT32>("Graphics", "MASKED_OC_OCCLUDER_AABB_SCREEN_SPACE_MIN_AREA");
+	mCullingSystem->mMaskedSWOcclusionCulling->mSolveMeshRoleStage.mOccluderAABBScreenSpaceMinArea = ConfigData::GetSingleton()->GetConfigData().GetValue<FLOAT64>("Graphics", "MASKED_OC_OCCLUDER_AABB_SCREEN_SPACE_MIN_AREA");
 
 	if(dooms::ui::MaskedOcclusionCulliingDebugger::GetSingleton() != nullptr)
 	{
